wget Fresh Install Cert Issue..fix
In one instance a fresh install left wget with no certs and so it yelled mightily at getting an https document. Much as I hate to put this in  using wget --no-check-certificate  gets around this. But please do not do this elsewhere unless you are really really and again really sure of your source.