[anyfesto-issue16] Clarify language around Anyfesto's discoverability
Summary
=======
Make clearer that Anyfesto's entry-point is a Wifi access point!

Details
=======
This PR results from having having known about the Anyfesto project for three gods-damned years and me not understanding that the thing appears as an access point to geo-adjacent users.

That fact should be right on the label.  "Captive Portal" didn't mean a thing to me until I looked it up. I doubt many people outside of IT will know wth it means either, it's too esoteric.

Discoverability and adoption of Anyfesto as a project might increase if it's made clearer in README.md how an Anyfest box is discoverable to the world.

Testing
=======
Maybe the fact of Anyfesto's Wifi-ness is obvious to everybody but me?  Cheers, anyhow...